modified script name
